Rename files_uploading to vcf_uploading
