+++ nationality_prediction/predictors.py
@@ -7,7 +7,7 @@ from django.core.files.uploadedfile import InMemoryUploadedFile
 from django.utils.translation import gettext_lazy as _
 from loguru import logger
 
-from files_uploading.vcf_processing import VCFFile
+from vcf_uploading.vcf_processing import VCFFile
 from nationality_prediction.command_line_tools import (run_fastngsadmix,
                                                        run_plink)
 from nationality_prediction.constants import (FAST_NGS_ADMIX_OUTPUT,
