+++ vcf_uploading/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class FilesUploadingConfig(AppConfig):
-    name = "files_uploading"
+    name = "vcf_uploading"
