+++ vcf_uploading/templates/snp_search.html
