+++ vcf_uploading/templates/vcf_list.html
